Fix H100 smoke crashes in the LingBot DICE-RL trainer.
Cast bf16 prior tensors to float32 before numpy and the residual MLPs, isolate critic VAE encodes from the AR cache, and compute n-step returns from successor chunks instead of bootstrapping on s.

+++ script/lingbot_rl_data.py
@@ -44,7 +44,7 @@ def expert_fingerprint(manifest, norm, episodes=None):
 
 def _tensor_state(state):
     if torch.is_tensor(state):
-        state = state.detach().cpu().numpy()
+        state = state.detach().float().cpu().numpy()
     state = np.asarray(state, dtype=np.float32)
     if state.ndim == 2:
         state = state[0]
